Added JavaDoc.  Slightly reformated some TopicService code

+++ src/services/TopicService.java
@@ -5,9 +5,12 @@ import java.rmi.RemoteException;
 import java.util.List;
 import java.util.UUID;
 
+import javax.naming.directory.InvalidAttributeValueException;
+
 import org.apache.commons.lang3.StringUtils;
 
 import exceptions.DuplicateEntryException;
+import exceptions.ResourceNotFoundException;
 import models.JMSTopic;
 import models.JMSTopicDeleted;
 import models.JMSTopicUser;
@@ -37,8 +40,15 @@ public class TopicService implements Serializable {
 	private EntryLookupHelper lookupHelper = new EntryLookupHelper();
 
 	private TopicService() {
+		// Uninstantiable singleton
 	}
 
+	/**
+	 * If a topic service instance exists, it is returned. If not, one is
+	 * created, then returned.
+	 * 
+	 * @return The TopicService instance
+	 */
 	public static TopicService getTopicService() {
 		if (topicService == null) {
 			topicService = new TopicService();
@@ -47,21 +57,36 @@ public class TopicService implements Serializable {
 		return topicService;
 	}
 
+	/**
+	 * Writes a topic to the JavaSpace
+	 * 
+	 * @param topic
+	 * @throws Exception
+	 */
 	public void createTopic(JMSTopic topic) throws Exception {
-		Transaction transaction = null;
+		Transaction transaction = TransactionHelper.getTransaction(3000);
 
 		try {
-			transaction = TransactionHelper.getTransaction(3000);
-
-			if (isValidTopic(topic) && !topicExistsInSpace(topic, transaction)) {
-				space.write(topic, transaction, Lease.FOREVER);
-				transaction.commit();
+			// If the topic is valid (non-null name, base name, and ID)
+			if (isValidTopic(topic)) {
+				// If the topic does not already exist in the space
+				if (!topicExistsInSpace(topic, transaction)) {
+					// Write it to the space and commit the transaction
+					space.write(topic, transaction, Lease.FOREVER);
+					transaction.commit();
+				} else {
+					// Otherwise it's a duplicate, so throw an exception
+					throw new DuplicateEntryException(
+							"Failed to create topic.  Topic baseName or id matches with an existing topic.");
+				}
 			} else {
-				throw new DuplicateEntryException(
-						"Failed to create topic.  Topic baseName or id matches with an existing topic.");
+				// Otherwise the Topic object is invalid, so throw an exception
+				throw new InvalidAttributeValueException("Topic being creates is invalid (one or more fields null)");
 			}
 
 		} catch (Exception e) {
+			// If anything in the try block throws an error, abort the
+			// transaction before rethrowing the error
 			if (transaction != null) {
 				try {
 					transaction.abort();
@@ -75,24 +100,47 @@ public class TopicService implements Serializable {
 		}
 	}
 
+	/**
+	 * Gets all of the topics in the space.
+	 * 
+	 * @return All of the topics in the space.
+	 */
 	public List<JMSTopic> getAllTopics() {
 		return lookupHelper.findAllMatchingTemplate(space, new JMSTopic());
 	}
 
-	public JMSTopic getTopicByName(String name) {
+	/**
+	 * Looks for a topic in the space with the same base name.
+	 * 
+	 * @param baseName
+	 *            The base name to search for.
+	 * @return If a matching topic exists, that topic is returned, otherwise
+	 *         <code>null</code> is returned.
+	 */
+	public JMSTopic getTopicByBaseName(String baseName) {
 		JMSTopic template = new JMSTopic();
-		template.setBaseName(name);
+		template.setBaseName(baseName);
 
-		JMSTopic topic = null;
+		JMSTopic topicFound = null;
 
 		try {
-			topic = (JMSTopic) space.readIfExists(template, null, 1000);
+			topicFound = (JMSTopic) space.readIfExists(template, null, 1000);
 		} catch (Exception e) {
+			System.err.println("Failed to get topic by base name. Base name queried: '" + baseName + "'");
 			e.printStackTrace();
 		}
-		return topic;
+
+		return topicFound;
 	}
 
+	/**
+	 * Looks for a topic in the space with the same UUID.
+	 * 
+	 * @param baseName
+	 *            The UUID to search for.
+	 * @return If a matching topic exists, that topic is returned, otherwise
+	 *         <code>null</code> is returned.
+	 */
 	public JMSTopic getTopicById(UUID id) {
 		JMSTopic template = new JMSTopic();
 		template.setId(id);
@@ -120,7 +168,7 @@ public class TopicService implements Serializable {
 	 *            The topic to check for existence of
 	 * 
 	 * @return Whether or not a topic exists in the space with all fields the
-	 *         same as specified in the paramater passed in. Also returns false
+	 *         same as specified in the parameter passed in. Also returns false
 	 *         if exception occurs
 	 */
 	public boolean doesTopicExistInSpace(JMSTopic topic, Transaction transaction) {
@@ -138,22 +186,40 @@ public class TopicService implements Serializable {
 	}
 
 	public void deleteTopic(JMSTopic topic) {
-		try {
-			space.takeIfExists(topic, null, 3000);
 
-			deleteAllTopicUsers(topic);
-			MessageService.getMessageService().deleteAllTopicMessages(topic);
-			
-			space.write(new JMSTopicDeleted(topic), null, 1000l * 60l);
-		} catch (Exception e) {
-			e.printStackTrace();
+		if (isValidTopic(topic)) {
+			try {
+				space.takeIfExists(topic, null, 3000);
+
+				deleteAllTopicUsers(topic);
+				MessageService.getMessageService().deleteAllTopicMessages(topic);
+
+				space.write(new JMSTopicDeleted(topic), null, 1000l * 60l);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		} else {
+			System.err.println("Attempted to delete topic with null fields.  "
+					+ "Due to how JavaSpaces work, this will delete one at random and is not allowed.");
 		}
 	}
 
+	/**
+	 * Gets all users that are currently in a given topic.
+	 * 
+	 * @param topic The topic to get users for.
+	 * 
+	 * @return All of the users in a given topic.
+	 */
 	public List<JMSTopicUser> getAllTopicUsers(JMSTopic topic) {
 		return lookupHelper.findAllMatchingTemplate(space, new JMSTopicUser(topic));
 	}
 
+	/**
+	 * Removes all users from a given topic.
+	 * 
+	 * @param topic A topic to remove all of the users from.
+	 */
 	private void deleteAllTopicUsers(JMSTopic topic) {
 		JMSTopicUser template = new JMSTopicUser(topic);
